refactor: separate CallvanFABDefaults and CallvanRouteState
