+++ feature/callvan/src/main/java/in/koreatech/koin/feature/callvan/enums/CallvanRouteState.kt
@@ -1,0 +1,9 @@
+package `in`.koreatech.koin.feature.callvan.enums
+
+enum class CallvanRouteState {
+    DEFAULT,
+    JOINED,
+    CLOSED,
+    OWNER_ACTIVE,
+    OWNER_CLOSED
+}
