+++ feature/callvan/src/main/java/in/koreatech/koin/feature/callvan/ui/list/component/CallvanFAB.kt
@@ -6,10 +6,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.WindowInsetsSides
-import androidx.compose.foundation.layout.only
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.systemBars
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.layout.wrapContentSize
@@ -29,15 +26,6 @@ import `in`.koreatech.koin.core.designsystem.noRippleClickable
 import `in`.koreatech.koin.core.designsystem.theme.KoinTheme
 import `in`.koreatech.koin.core.designsystem.theme.RebrandKoinTheme
 import `in`.koreatech.koin.feature.callvan.R
-
-object CallvanFABDefaults {
-    val windowInsets: WindowInsets
-        @Composable
-        get() =
-            WindowInsets.systemBars.only(
-                WindowInsetsSides.Horizontal + WindowInsetsSides.Bottom
-            )
-}
 
 @Composable
 fun CallvanFAB(
