+++ feature/callvan/src/main/java/in/koreatech/koin/feature/callvan/ui/list/component/CallvanFABDefaults.kt
@@ -1,0 +1,16 @@
+package `in`.koreatech.koin.feature.callvan.ui.list.component
+
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.WindowInsetsSides
+import androidx.compose.foundation.layout.only
+import androidx.compose.foundation.layout.systemBars
+import androidx.compose.runtime.Composable
+
+object CallvanFABDefaults {
+    val windowInsets: WindowInsets
+        @Composable
+        get() =
+            WindowInsets.systemBars.only(
+                WindowInsetsSides.Horizontal + WindowInsetsSides.Bottom
+            )
+}
