+++ feature/callvan/src/main/java/in/koreatech/koin/feature/callvan/ui/list/component/CallvanListItem.kt
@@ -14,19 +14,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import `in`.koreatech.koin.core.designsystem.theme.KoinTheme
 import `in`.koreatech.koin.core.designsystem.theme.RebrandKoinTheme
+import `in`.koreatech.koin.feature.callvan.enums.CallvanRouteState
 import `in`.koreatech.koin.feature.callvan.ui.component.CallvanRouteCount
 import `in`.koreatech.koin.feature.callvan.ui.component.CallvanRouteDate
 import `in`.koreatech.koin.feature.callvan.ui.component.CallvanRouteInfo
 import `in`.koreatech.koin.feature.callvan.ui.list.model.CallvanListItemClickListener
 import `in`.koreatech.koin.feature.callvan.ui.list.model.CallvanListUiState
 
-enum class CallvanRouteState {
-    DEFAULT,
-    JOINED,
-    CLOSED,
-    OWNER_ACTIVE,
-    OWNER_CLOSED
-}
 
 @Composable
 fun CallvanListItem(
